feat: extract to components
