+++ app/routes/_layout.studio.tsx
@@ -5,11 +5,13 @@ import { SpotifyWebApi } from "@thomasngrlt/spotify-web-api-ts";
 import type { Artist } from "@thomasngrlt/spotify-web-api-ts/types/types/SpotifyObjects";
 import { Suspense, useState } from "react";
 import { z } from "zod";
-import { motion } from "framer-motion";
 
+import { ArtistChip } from "~/components/ArtistChip";
 import { Label } from "~/components/Label";
 import { NumberInput } from "~/components/NumberInput";
 import { RangeSlider } from "~/components/Range";
+import { RangeGroup } from "~/components/RangeGroup";
+import { Spinner } from "~/components/Spinner";
 import { TextInput } from "~/components/TextInput";
 import { getAllFollowedArtists } from "~/models/api.server";
 import { getSession } from "~/sessions";
@@ -146,7 +148,7 @@ export default function StudioPage() {
 	};
 
 	const ArtistList = ({ artists }: { artists: Artist[] }) => (
-		<>
+		<div className="flex max-h-full w-full flex-wrap items-start justify-center gap-2">
 			{artists.map(({ name, images, id }) => (
 				<ArtistChip
 					key={id}
@@ -156,7 +158,7 @@ export default function StudioPage() {
 					selected={Boolean(selectedArtists.find((artist) => artist.id === id))}
 				/>
 			))}
-		</>
+		</div>
 	);
 
 	return (
@@ -208,7 +210,7 @@ export default function StudioPage() {
 						id="tempo"
 						min={30}
 						max={300}
-						className="w-[350px]"
+						className="w-full max-w-[350px]"
 					/>
 				</RangeGroup>
 				<RangeGroup
@@ -221,7 +223,7 @@ export default function StudioPage() {
 						min={0}
 						max={100}
 						step={1}
-						className="w-[350px]"
+						className="w-full max-w-[350px]"
 					/>
 				</RangeGroup>
 				<RangeGroup
@@ -234,90 +236,11 @@ export default function StudioPage() {
 						min={0}
 						max={1}
 						step={0.01}
-						className="w-[350px]"
+						className="w-full max-w-[350px]"
 					/>
 				</RangeGroup>
 				<button type="submit">Create</button>
 			</Form>
 		</div>
-	);
-}
-
-type RangeGroupProps = {
-	label: string;
-	leftText: string;
-	rightText: string;
-	children: React.ReactNode;
-};
-
-function RangeGroup({ label, leftText, rightText, children }: RangeGroupProps) {
-	return (
-		<div>
-			<Label>{label}</Label>
-			<div>
-				{children}
-				<div className="flex justify-between">
-					<span className="text-xs text-neutral-400">{leftText}</span>
-					<span className="text-xs text-neutral-400">{rightText}</span>
-				</div>
-			</div>
-		</div>
-	);
-}
-
-type ArtistChipProps = {
-	image: string;
-	text: string;
-	selected: boolean;
-	onClick: () => void;
-};
-
-function ArtistChip({ image, text, onClick, selected }: ArtistChipProps) {
-	return (
-		<div
-			onClick={onClick}
-			className={`flex h-8 w-max items-center gap-x-2 rounded-full border-2 bg-neutral-800 py-1 pl-1 pr-2 transition-colors hover:cursor-pointer hover:bg-neutral-700
-			${
-				selected
-					? "border-green-600 hover:border-green-500"
-					: "border-neutral-600 hover:border-neutral-500"
-			}
-			`}>
-			<img
-				src={image}
-				alt={`${text} icon`}
-				width={20}
-				height={20}
-				className="rounded-full"
-			/>
-			<p
-				className={`text-sm transition-colors  ${
-					selected ? "text-white" : "text-neutral-400"
-				}`}>
-				{text}
-			</p>
-		</div>
-	);
-}
-
-function Spinner() {
-	return (
-		<motion.svg
-			fill="#ffffff"
-			width="32px"
-			height="32px"
-			viewBox="0 0 32 32"
-			version="1.1"
-			xmlns="http://www.w3.org/2000/svg"
-			animate={{ rotate: 360 }}
-			transition={{
-				repeat: Infinity,
-				bounce: 0,
-				ease: "linear",
-				duration: 0.75,
-			}}>
-			<title>spinner-one-third</title>
-			<path d="M16 0.75c-0.69 0-1.25 0.56-1.25 1.25s0.56 1.25 1.25 1.25v0c7.042 0.001 12.75 5.71 12.75 12.751 0 3.521-1.427 6.709-3.734 9.016v0c-0.226 0.226-0.365 0.538-0.365 0.883 0 0.69 0.56 1.25 1.25 1.25 0.346 0 0.659-0.14 0.885-0.367l0-0c2.759-2.76 4.465-6.572 4.465-10.782 0-8.423-6.828-15.251-15.25-15.251h-0z" />
-		</motion.svg>
 	);
 }
